+++ .mega-linter.yml
@@ -11,5 +11,6 @@ DISABLE:
 SHOW_ELAPSED_TIME: true
 FILEIO_REPORTER: false
 EXCLUDED_DIRECTORIES:
-  - megalinter-reports/
+  - megalinter-reports
+  - node_modules
 # DISABLE_ERRORS: true # Uncomment if you want MegaLinter to detect errors but not block CI to pass
