+++ package.json
@@ -28,13 +28,15 @@
     "dayjs": "^1.11.9",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
-    "express-handlebars": "^6.0.1",
+    "express-handlebars": "^7.1.2",
     "mongoose": "^6.0.13",
-    "mongoose-lean-virtuals": "^0.9.1",
-    "multer": "^1.4.3"
+    "mongoose-lean-virtuals": "^0.9.1"
   },
   "devDependencies": {
     "browser-sync": "^2.29.1",
-    "nodemon": "^2.0.22"
+    "nodemon": "^3.0.1",
+    "@types/body-parser": "^1.19.2",
+    "@types/express": "^4.17.17",
+    "tsc": "^2.0.4"
   }
 }
