+++ src/config/bs-config.js
