+++ src/scripts/defaultLoanerCreation.js
