+++ src/views/error/404.hbs
