+++ src/views/layouts/main.hbs
@@ -20,7 +20,7 @@
             <img src="/static/washulogo.svg" width="321" height="28" alt="washu logo">
         </div>
     </div>
-    <div>
+    <div class="container">
         {{{body}}}
     </div>
