+++ src/views/record.hbs
